refactor(billing): drop BillingPdfService shim, import pdf/ directly
The re-export shim was only bridging the split. Repoint all importers
(both orchestrators + their tests, domain context.ts and index.ts) to
./services/pdf, move the test into pdf/index.test.ts, and delete the shim.
88 tests pass; tsc + biome clean.
